Fix incorrect use of mutexes.
This commit fixes the two incorrect (double-release) issues
demonstrated by the previous commit.

All we strictly needed to do was to avoid the double-release in
`client.h::WriteMaybeLast()` and `server.h::WriteMaybeFinish()`.
However, we're taking this opportunity to improve how we use
mutexes, by introducing `absl::MutexLock` and
`absl::ReleasableMutexLock`, which ensure that we can never
forget to release a Mutex. See [here](https://abseil.io/docs/cpp/guides/synchronization#the-mutexlock-wrapper)
for more info.

For bonus points we also demonstrate the use of
[Thread Annotations](https://abseil.io/docs/cpp/guides/synchronization#thread-annotations)
which further improve our ability to find any future mistakes.