Bump version of stout-eventuals, fix local build
Update stout-eventuals in light of recent changes, in particular the
inclusion of Post and Get that relies on curl. The inclusion of curl
introduces two new repo dependencies included in this commit.

As outlined in issue #17, building `:grpc` locally using `gcc` is
currently broken by `.bazelrc`. This commit works around it by removing
the offending compiler flags, allowing the library to be built locally.
This fixes local compilation but it does not fix the issue: we want
those compiler flags in there where possible, i.e., for `clang`.