Added check_format_code.yml
This .yml file helps to verify that a repo's formatting matches our
specifications (`.clang-format` file in the root directory), but doesn't
automatically fix any issues.

For now I need to figure out how to make this action automatically
check if every line of the code exceeds 80 characters and make this
action work with BUILD.bazel and WORKSPACE.bazel files.